Merged PR 2742: Restore phone number tap analytics while also not crashing for VoiceOver users
A pretty ingenious solution if I do say so myself!

Background: During R1 or R2, we discovered that a UITextView that had implemented the `shouldInteractWithURL` delegate method would cause a crash when a VoiceOver user tapped a data detected link (such as a phone number). Since that delegate method is the only way to be notified of these taps, we decided to remove the tap analytics events on iOS so that we did not have to implement the delegate, and the app would not crash.

With this change, we conditionally become the text view's delegate only if VoiceOver is not running, and forward the  `shouldInteractWithURL` callback to our own custom delegate method. As such, a non-VoiceOver user tapping a phone number link will now accurately trigger the respective analytic event. The trade off is that a VoiceOver user tapping a phone number will not send analytics, but the accessibility will be maintained, the link is usable, and the app will not crash.

Related work items: #15278